+++ apps/client/src/pages/ai-results/apis/queries.ts
@@ -1,10 +1,7 @@
 import { useInfiniteQuery, useQuery } from '@tanstack/react-query';
 
 import { api } from '@shared/apis/instance';
-import {
-  mapApiMemoToMockMemo,
-  MockMemo,
-} from '@shared/components/memo-list-view/components/memo-list/types/memo';
+import { mapApiMemoToMockMemo, MockMemo } from '@shared/types/memo';
 import { components } from '@shared/types/schema';
 
 import { AI_END_POINT } from './end-point';
