+++ apps/client/src/pages/new-memo/components/input-content/input-content.css.ts
