+++ apps/client/src/shared/components/ai-prompt/types/types.ts
