--- apps/client/src/shared/components/memo-editor/index.ts
@@ -1,3 +1,0 @@
-export { default as InputContent } from './input-content';
-export { default as InputTitle } from './input-title';
-export { default as LabelSelect } from './label-select';
