--- apps/client/src/shared/components/memo-list-view/components/memo-list/index.ts
@@ -1,1 +1,0 @@
-export { default as MemoCardGrid } from './memo-card-grid';
