--- apps/client/src/shared/components/memo-list-view/components/tree-view/index.ts
@@ -1,1 +1,0 @@
-export { default as TreeView } from './tree-view';
